redis_client.py
import os
import redis
from dotenv import load_dotenv

load_dotenv()

REDIS_URL = os.getenv("REDIS_URL")
REDIS_TTL = int(os.getenv("REDIS_TTL", 1800))

redis_client = redis.from_url(REDIS_URL, decode_responses=True)
